Move logic into own module
